+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## v3.1.15
+## v3.2.0
 
 <!--Releasenotes start-->
 - Fixed buttons on the bottom right of videos on the video page being shifted right into the "More Videos" section if a shuffle button is present.
 - Fixed an issue where private or deleted videos could be chosen when shuffling.
+- Fixed videos being removed from the extension's saved channel data if a connection problem prevented the extension from checking whether they can still be watched.
+- The extension will now show a notice if the availability of multiple videos in a row cannot be checked.
 <!--Releasenotes end-->
 
 ## v3.1.14